pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easypost-tools"
description = "A large collection of supplemental tools for the EasyPost API spanning various programming languages."
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Justintime50" }]
urls = { Homepage = "http://github.com/justintime50/easypost-tools" }
dependencies = [
    "beautifulsoup4 ==4.*",
    "easypost ==10.*",
    "flask ==3.*",
    "Pillow ==12.*",
    "PyPDF2 ==1.27.*",
    "python-dotenv ==1.*",
    "pyzbar ==0.1.*",
]
optional-dependencies = { dev = ["ruff == 0.14.*"] }

[tool.setuptools.packages.find]
exclude = ["examples", "test"]

[tool.ruff]
line-length = 120
